Review and improvements.

+++ app/src/main/java/de/xikolo/model/DownloadModel.java
@@ -1,8 +1,10 @@
 package de.xikolo.model;
 
+import android.app.Activity;
 import android.app.DownloadManager;
 import android.os.Environment;
 import android.util.Log;
+import android.widget.Toast;
 
 import com.path.android.jobqueue.JobManager;
 
@@ -11,6 +13,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
+import de.greenrobot.event.EventBus;
 import de.xikolo.GlobalApplication;
 import de.xikolo.R;
 import de.xikolo.data.entities.Course;
@@ -18,6 +21,8 @@ import de.xikolo.data.entities.Download;
 import de.xikolo.data.entities.Item;
 import de.xikolo.data.entities.Module;
 import de.xikolo.data.net.DownloadHelper;
+import de.xikolo.model.events.PermissionDeniedEvent;
+import de.xikolo.model.events.PermissionGrantedEvent;
 import de.xikolo.model.jobs.RetrieveContentLengthJob;
 import de.xikolo.util.Config;
 import de.xikolo.util.ExternalStorageUtil;
@@ -27,8 +32,59 @@ public class DownloadModel extends BaseModel {
 
     public static final String TAG = DownloadModel.class.getSimpleName();
 
-    public DownloadModel(JobManager jobManager) {
+    private Activity activity;
+
+    private PermissionsModel permissionsModel;
+
+    public enum PendingAction {
+            START, DELETE, CANCEL;
+
+        private String uri;
+        private DownloadFileType type;
+        private Course course;
+        private Module module;
+        private Item item;
+
+        public void savePayload(String uri, DownloadFileType type, Course course, Module module, Item item) {
+            this.uri = uri;
+            this.type = type;
+            this.course = course;
+            this.module = module;
+            this.item = item;
+        }
+
+        public String getUri() {
+            return uri;
+        }
+
+        public DownloadFileType getType() {
+            return type;
+        }
+
+        public Course getCourse() {
+            return course;
+        }
+
+        public Module getModule() {
+            return module;
+        }
+
+        public Item getItem() {
+            return item;
+        }
+    }
+
+    private PendingAction pendingAction;
+
+    public DownloadModel(JobManager jobManager, Activity activity) {
         super(jobManager);
+        this.activity = activity;
+
+        this.permissionsModel = new PermissionsModel(jobManager, activity);
+
+        this.pendingAction = null;
+
+        EventBus.getDefault().register(this);
     }
 
     public void getRemoteDownloadFileSize(Result<Long> result, String url) {
@@ -40,18 +96,23 @@ public class DownloadModel extends BaseModel {
             Log.d(TAG, "Start download for " + uri);
         }
         if (ExternalStorageUtil.isExternalStorageWritable()) {
-            String file = item.title + type.getFileSuffix();
-            String filename = buildDownloadFilename(type, course, module, item);
+            if (permissionsModel.requestPermission(PermissionsModel.WRITE_EXTERNAL_STORAGE) == 1) {
+                String file = item.title + type.getFileSuffix();
+                String filename = buildDownloadFilename(type, course, module, item);
 
-            if (downloadExists(filename)) {
-                ToastUtil.show(GlobalApplication.getInstance(), R.string.toast_file_already_downloaded);
+                if (downloadExists(filename)) {
+                    ToastUtil.show(GlobalApplication.getInstance(), R.string.toast_file_already_downloaded);
+                } else {
+
+                    File dlFile = new File(filename);
+
+                    createFolderIfNotExists(new File(dlFile.getAbsolutePath().replace(file, "")));
+
+                    return DownloadHelper.request(uri, "file://" + dlFile.getAbsolutePath(), file);
+                }
             } else {
-
-                File dlFile = new File(filename);
-
-                createFolderIfNotExists(new File(dlFile.getAbsolutePath().replace(file, "")));
-
-                return DownloadHelper.request(uri, "file://" + dlFile.getAbsolutePath(), file);
+                pendingAction = PendingAction.START;
+                pendingAction.savePayload(uri, type, course, module, item);
             }
         } else {
             Log.w(TAG, "No write access for external storage");
@@ -62,17 +123,23 @@ public class DownloadModel extends BaseModel {
 
     public boolean deleteDownload(DownloadFileType type, Course course, Module module, Item item) {
         if (ExternalStorageUtil.isExternalStorageWritable()) {
-            String filename = buildDownloadFilename(type, course, module, item);
+            if (permissionsModel.requestPermission(PermissionsModel.WRITE_EXTERNAL_STORAGE) == 1) {
+                String filename = buildDownloadFilename(type, course, module, item);
 
-            if (Config.DEBUG) {
-                Log.d(TAG, "Delete download " + filename);
-            }
+                if (Config.DEBUG) {
+                    Log.d(TAG, "Delete download " + filename);
+                }
 
-            if (!downloadExists(filename)) {
-                return false;
+                if (!downloadExists(filename)) {
+                    return false;
+                } else {
+                    File dlFile = new File(filename);
+                    return dlFile.delete();
+                }
             } else {
-                File dlFile = new File(filename);
-                return dlFile.delete();
+                pendingAction = PendingAction.DELETE;
+                pendingAction.savePayload(null, type, course, module, item);
+                return false;
             }
         } else {
             Log.w(TAG, "No write access for external storage");
@@ -83,26 +150,57 @@ public class DownloadModel extends BaseModel {
 
     public void cancelDownload(DownloadFileType type, Course course, Module module, Item item) {
         if (ExternalStorageUtil.isExternalStorageWritable()) {
-            String filename = buildDownloadFilename(type, course, module, item);
-            Download dl = new Download();
-            dl.localFilename = filename;
+            if (permissionsModel.requestPermission(PermissionsModel.WRITE_EXTERNAL_STORAGE) == 1) {
+                String filename = buildDownloadFilename(type, course, module, item);
+                Download dl = new Download();
+                dl.localFilename = filename;
 
-            if (Config.DEBUG) {
-                Log.d(TAG, "Cancel download " + filename);
-            }
-
-            Set<Download> dlSet = DownloadHelper.getAllDownloads();
-            for (Download download : dlSet) {
-                if (download.equals(dl)) {
-                    DownloadHelper.remove(download.id);
+                if (Config.DEBUG) {
+                    Log.d(TAG, "Cancel download " + filename);
                 }
-            }
 
-            deleteDownload(type, course, module, item);
+                Set<Download> dlSet = DownloadHelper.getAllDownloads();
+                for (Download download : dlSet) {
+                    if (download.equals(dl)) {
+                        DownloadHelper.remove(download.id);
+                    }
+                }
+
+                deleteDownload(type, course, module, item);
+            } else {
+                pendingAction = PendingAction.CANCEL;
+                pendingAction.savePayload(null, type, course, module, item);
+            }
         } else {
             Log.w(TAG, "No write access for external storage");
             ToastUtil.show(GlobalApplication.getInstance(), R.string.toast_no_external_write_access);
         }
+    }
+
+    public void onEvent(PermissionGrantedEvent permissionGrantedEvent) {
+        if (pendingAction != null) {
+            switch (pendingAction) {
+                case START:
+                    startDownload(pendingAction.getUri(), pendingAction.getType(),
+                            pendingAction.getCourse(), pendingAction.getModule(), pendingAction.getItem());
+                    break;
+                case DELETE:
+                    deleteDownload(pendingAction.getType(),
+                            pendingAction.getCourse(), pendingAction.getModule(), pendingAction.getItem());
+                    break;
+                case CANCEL:
+                    cancelDownload(pendingAction.getType(),
+                            pendingAction.getCourse(), pendingAction.getModule(), pendingAction.getItem());
+                    break;
+                default:
+                    pendingAction = null;
+            }
+        }
+        pendingAction = null;
+    }
+
+    public void onEvent(PermissionDeniedEvent permissionDeniedEvent){
+        pendingAction = null;
     }
 
     public Download getDownload(DownloadFileType type, Course course, Module module, Item item) {
